quotegen/requirements.txt
uvicorn
daiquiri
python-multipart
python-dotenv
requests
six
cryptography
fastapi[all]
sqlalchemy
pydantic
